[TensorExpr] Explicitly cast to bool results of comparison ops in kernel.cpp.
Previously, we've been using operators <, >, ==, et al. and relied on
the dtype to be picked automatically. It led to a wrong dtype being
picked for the result, but that choice was overwritten by the type
explicitly specified in JIT IR, which we were lowering. Now we are
moving towards using shape inference instead of relying on all types
being specified in the IR, and that made this issue to immediately pop
up.

ghstack-source-id: 5d0c7d778d7ef69c0471c070bc013bb828abdef8
Pull Request resolved: https://github.com/pytorch/pytorch/pull/42201