Moving ProcessGroupNCCLTest to Gtest
ProcessGroupNCCLTest is currently written without any testing framework, and all tests are simply called from the main function and throw exceptions upon failure. As a result, it is hard to debug and pinpoint which tests have succeeded/failed.

This PR moves ProcessGroupNCCLTest to gtest with appropriate setup and skipping functionality in the test superclass.

Differential Revision: [D22782673](https://our.internmc.facebook.com/intern/diff/D22782673/)

ghstack-source-id: 108720394
Pull Request resolved: https://github.com/pytorch/pytorch/pull/42208