[pytorch][ci] install nightly instead of stable libtorch for mobile CIs
Summary:
Mobile custom build CI jobs need desktop version libtorch to prepare
models and dump root ops.

Ideally we should use the libtorch built on the PR so that backward
incompatible changes won't break this script - but it will significantly
slow down mobile CI jobs.

This PR changed it to install nightly instead of stable so that we have
an option to temporarily skip mobile CI jobs on BC-breaking PRs until
they are in nightly.

ghstack-source-id: d049fd66c673a38ade68d4f1fac751d77b7c655d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/42220